requirements.txt
setuptools==15
tornado==4.1
psycopg2==2.6
sqlalchemy>=0.9,<1.1
netifaces==0.10.4
pycrypto==2.6.1
pytz==2015.4
psutil>=0.6,<4.3
six==1.9
requests==2.7
gevent==1.0.2
werkzeug==0.10.4
patool==1.7
bcrypt==2.0
chardet==2.3
backports.csv==1.0.2 # not needed in Python3

# Only for some importers:
pyyaml==3.11

# Only for printing:
pycups==1.9.72
PyPDF2==1.24
